+++ changelog.markdown
@@ -1,3 +1,16 @@
+# 2.0.0 Arena for the Net
+
+- Explicit support for `<textarea>` elements
+- Added `autoShowOnUpDown` option
+
+# 1.2.0 Exposé
+
+- Exposed `defaultRenderer` method
+- Exposed `defaultGetText` method
+- Exposed `defaultGetValue` method
+- Exposed `defaultSetter` method
+- Exposed `defaultFilter` method
+
 # 1.1.1 Cross Polinize
 
 - Updated `crossvent` to `1.3.1`